moved permission check code into member.helpers, beefed up permission-checking for editing bands
